remove model_ids from agent_base and model_base
