+++ parl/core/fluid/tests/agent_model_ids_test.py
@@ -13,11 +13,10 @@
 # limitations under the License.
 
 import unittest
-from parl.core.model_base import ModelBase
-from parl.core.algorithm_base import AlgorithmBase
+import parl
 
 
-class MockModel(ModelBase):
+class MockModel(parl.Model):
     def __init__(self, weights, model_id=None):
         super(MockModel, self).__init__(model_id)
         self.weights = weights
@@ -29,7 +28,7 @@ class MockModel(ModelBase):
         self.weights = weights
 
 
-class TestAlgorithm(AlgorithmBase):
+class TestAlgorithm(parl.Algorithm):
     def __init__(self):
         self.model1 = MockModel(1)
         self.model2 = MockModel(2)
@@ -39,7 +38,7 @@ class TestAlgorithm(AlgorithmBase):
         self.model_dict2 = {'k1': MockModel(7), 'k2': MockModel(8)}
 
 
-class TestAlgorithm2(AlgorithmBase):
+class TestAlgorithm2(parl.Algorithm):
     def __init__(self):
         self.model1 = MockModel(1, model_id='id1')
         self.model2 = MockModel(2, model_id='id2')
@@ -55,13 +54,20 @@ class TestAlgorithm2(AlgorithmBase):
         }
 
 
-class AlgorithmBaseTest(unittest.TestCase):
+class TestAgent(parl.Agent):
+    def build_program(self):
+        pass
+
+
+class AgentBaseTest(unittest.TestCase):
     def setUp(self):
-        self.alg1 = TestAlgorithm()
-        self.alg2 = TestAlgorithm()
+        alg1 = TestAlgorithm()
+        alg2 = TestAlgorithm()
+        self.agent1 = TestAgent(alg1)
+        self.agent2 = TestAgent(alg2)
 
     def test_get_weights(self):
-        weights = self.alg1.get_weights()
+        weights = self.agent1.get_weights()
         expected_dict = {
             'model1': 1,
             'model2': 2,
@@ -91,79 +97,10 @@ class AlgorithmBaseTest(unittest.TestCase):
                 'k2': -8
             }
         }
-        self.alg1.set_weights(expected_dict)
-        self.assertDictEqual(self.alg1.get_weights(), expected_dict)
+        self.agent1.set_weights(expected_dict)
+        self.assertDictEqual(self.agent1.get_weights(), expected_dict)
 
-    def test_set_weights_with_inconsistent_weights_case1(self):
-        inconsistent_weights = {
-            'model0': -1,
-            'model2': -2,
-            'model_list1': [-3],
-            'model_list2': [-4, -5],
-            'model_dict1': {
-                'k1': -6
-            },
-            'model_dict2': {
-                'k1': -7,
-                'k2': -8
-            }
-        }
-        with self.assertRaises(AssertionError):
-            self.alg1.set_weights(inconsistent_weights)
-
-    def test_set_weights_with_inconsistent_weights_case2(self):
-        inconsistent_weights = {
-            'model1': -1,
-            'model2': -2,
-            'model_list1': [-3],
-            'model_list2': [-4],
-            'model_dict1': {
-                'k1': -6
-            },
-            'model_dict2': {
-                'k1': -7,
-                'k2': -8
-            }
-        }
-        with self.assertRaises(AssertionError):
-            self.alg1.set_weights(inconsistent_weights)
-
-    def test_set_weights_with_inconsistent_weights_case3(self):
-        inconsistent_weights = {
-            'model1': -1,
-            'model2': -2,
-            'model_list1': [-3],
-            'model_list2': [-4, -5],
-            'model_dict1': {
-                'k1': -6
-            },
-            'model_dict2': {
-                'k1': -7,
-            }
-        }
-        with self.assertRaises(AssertionError):
-            self.alg1.set_weights(inconsistent_weights)
-
-    def test_set_weights_with_redundant_weights(self):
-        redundant_weights = {
-            'model0': 0,
-            'model1': -1,
-            'model2': -2,
-            'model_list1': [-3],
-            'model_list2': [-4, -5],
-            'model_list3': [44, 55],
-            'model_dict1': {
-                'k1': -6
-            },
-            'model_dict2': {
-                'k1': -7,
-                'k2': -8
-            },
-            'model_dict3': {
-                'k1': -77,
-                'k2': -88
-            }
-        }
+    def test_get_and_set_weights_between_agents(self):
         expected_dict = {
             'model1': -1,
             'model2': -2,
@@ -177,38 +114,22 @@ class AlgorithmBaseTest(unittest.TestCase):
                 'k2': -8
             }
         }
-        self.alg1.set_weights(expected_dict)
-        self.assertDictEqual(self.alg1.get_weights(), expected_dict)
+        self.agent1.set_weights(expected_dict)
+        new_weights = self.agent1.get_weights()
 
-    def test_get_and_set_weights_between_algorithms(self):
-        expected_dict = {
-            'model1': -1,
-            'model2': -2,
-            'model_list1': [-3],
-            'model_list2': [-4, -5],
-            'model_dict1': {
-                'k1': -6
-            },
-            'model_dict2': {
-                'k1': -7,
-                'k2': -8
-            }
-        }
-        self.alg1.set_weights(expected_dict)
-        new_weights = self.alg1.get_weights()
-
-        self.alg2.set_weights(new_weights)
-        self.assertDictEqual(self.alg2.get_weights(), expected_dict)
+        self.agent2.set_weights(new_weights)
+        self.assertDictEqual(self.agent2.get_weights(), expected_dict)
 
     def test_get_model_ids(self):
         alg = TestAlgorithm2()
+        agent = TestAgent(alg)
         expected_model_ids = set(['id{}'.format(i + 1) for i in range(8)])
-        self.assertSetEqual(expected_model_ids, alg.get_model_ids())
+        self.assertSetEqual(expected_model_ids, agent.get_model_ids())
 
     def test_get_weights_with_model_ids(self):
-        weights = self.alg1.get_weights(model_ids=[
-            self.alg1.model1.model_id, self.alg1.model_list2[0].model_id, self.
-            alg1.model_dict2['k1'].model_id
+        weights = self.agent1.get_weights(model_ids=[
+            self.agent1.alg.model1.model_id, self.agent1.alg.model_list2[0].
+            model_id, self.agent1.alg.model_dict2['k1'].model_id
         ])
         expected_dict = {
             'model1': 1,
@@ -241,22 +162,25 @@ class AlgorithmBaseTest(unittest.TestCase):
             }
         }
 
-        self.alg1.set_weights(
+        self.agent1.set_weights(
             new_weights,
             model_ids=[
-                self.alg1.model1.model_id, self.alg1.model_list2[0].model_id,
-                self.alg1.model_dict2['k1'].model_id
+                self.agent1.alg.model1.model_id,
+                self.agent1.alg.model_list2[0].model_id,
+                self.agent1.alg.model_dict2['k1'].model_id
             ])
-        self.assertDictEqual(self.alg1.get_weights(), expected_dict)
+        self.assertDictEqual(self.agent1.get_weights(), expected_dict)
 
-    def test_get_and_set_weights_between_algorithms_with_model_ids(self):
-        alg1_model_ids = [
-            self.alg1.model1.model_id, self.alg1.model_list2[0].model_id,
-            self.alg1.model_dict2['k1'].model_id
+    def test_get_and_set_weights_between_agents_with_model_ids(self):
+        agent1_model_ids = [
+            self.agent1.alg.model1.model_id,
+            self.agent1.alg.model_list2[0].model_id,
+            self.agent1.alg.model_dict2['k1'].model_id
         ]
-        alg2_model_ids = [
-            self.alg2.model1.model_id, self.alg2.model_list2[0].model_id,
-            self.alg2.model_dict2['k1'].model_id
+        agent2_model_ids = [
+            self.agent2.alg.model1.model_id,
+            self.agent2.alg.model_list2[0].model_id,
+            self.agent2.alg.model_dict2['k1'].model_id
         ]
         new_weights = {
             'model1': -1,
@@ -278,11 +202,11 @@ class AlgorithmBaseTest(unittest.TestCase):
                 'k2': 8
             }
         }
-        self.alg1.set_weights(new_weights, alg1_model_ids)
-        alg1_weights = self.alg1.get_weights(alg1_model_ids)
+        self.agent1.set_weights(new_weights, agent1_model_ids)
+        agent1_weights = self.agent1.get_weights(agent1_model_ids)
 
-        self.alg2.set_weights(alg1_weights, alg2_model_ids)
-        self.assertDictEqual(self.alg2.get_weights(), expected_dict)
+        self.agent2.set_weights(agent1_weights, agent2_model_ids)
+        self.assertDictEqual(self.agent2.get_weights(), expected_dict)
 
 
 if __name__ == '__main__':
